[JENKINS-33520] provide a better way to restart Jenkins.
When run on windows locally Jenkins is not restartable - which causes many
tests to be skipped.  However for the cases where Jenkins can not restart
itself the Controller may be able to restart it (by killing the process
and then starting it again), so add a new helper and link this in so we
can get better test coverage on Windows platforms.